feat: Added the calculation of area of the triangles
Did this by retrieving the points and stacking the list of arrays into
one 3x3 numpy array. Then i removed the z-axis from and turned it into a
3x2 array. Then i used the formula for finging the area of a triangle
from three points.

IMPORTANT! Remember to implement this optimilazation when adding the
points to the triangle immidiatly, not when calculating the area.