feat: finished all initial calculations
The plot seems like it is working, which means the initial oil values
and the cell id's should be working correctly. Hopefully the same is
correct for the other mathematical functions.

Have finished computing everything i believe is necessary in order to
update oil values.

Only missing making the simulation.